+++ packages/__vital__/src/styleguide/Page/StyleGuideTemplate/index.tsx
@@ -19,9 +19,12 @@ import { Header } from './Header';
 import { FlowContainer } from './FlowContainer';
 import { Images } from './Images';
 import { Footer } from './Footer';
+import { Card } from './Card';
 import { Table } from './Table';
 import { List } from './List';
+import { Buttons } from './Buttons';
 import { _default } from './_default';
+import { Video } from './Video';
 
 export const __vital__StyleGuideTemplate = {
   _default,
@@ -33,6 +36,9 @@ export const __vital__StyleGuideTemplate = {
   Images,
   FlowContainer,
   Footer,
+  Video,
+  Card,
   Table,
   List,
+  Buttons,
 };