+++ packages/vital-elements/src/components/Typography/tokens/vitalTypography.ts
@@ -26,9 +26,6 @@ const meta = {
 };
 
 const Link = asElementToken({
-  Core: {
-    _: vitalFontSize.Base,
-  },
   Theme: {
     _: as(
       vitalTextDecoration.Bold,
@@ -118,6 +115,61 @@ const H5 = asElementToken({
   Meta: meta,
 });
 
+const HeaderLink = as(
+  vitalColor.TextPrimaryInteractive,
+  vitalTextDecoration.Underline,
+);
+
+const H1Link = asElementToken({
+  ...H1,
+  Theme: {
+    _: as(
+      vitalTextDecoration.Bold,
+      HeaderLink,
+    ),
+  },
+});
+
+const H2Link = asElementToken({
+  ...H2,
+  Theme: {
+    _: as(
+      vitalTextDecoration.Bold,
+      HeaderLink,
+    ),
+  },
+});
+
+const H3Link = asElementToken({
+  ...H3,
+  Theme: {
+    _: as(
+      vitalTextDecoration.Medium,
+      HeaderLink,
+    ),
+  },
+});
+
+const H4Link = asElementToken({
+  ...H4,
+  Theme: {
+    _: as(
+      vitalTextDecoration.Normal,
+      HeaderLink,
+    ),
+  },
+});
+
+const H5Link = asElementToken({
+  ...H5,
+  Theme: {
+    _: as(
+      vitalTextDecoration.Medium,
+      HeaderLink,
+    ),
+  },
+});
+
 const Body = asElementToken({
   Core: {
     _: vitalFontSize.Base,
@@ -165,6 +217,18 @@ const Rest = asElementToken({
   Meta: meta,
 });
 
+// Variant testing autoprefixer
+const Gradient = asElementToken({
+  Core: {
+    _: vitalFontSize.L,
+  },
+  Theme: {
+    _: as('bg-clip-text text-transparent bg-gradient-to-r',
+      'from-vital-secondary-eyebrow via-vital-primary-interactive to-vital-secondary-eyebrow',),
+  },
+  Meta: meta,
+});
+
 export default {
   Link,
   H1,
@@ -172,7 +236,13 @@ export default {
   H3,
   H4,
   H5,
+  H1Link,
+  H2Link,
+  H3Link,
+  H4Link,
+  H5Link,
   Body,
   Eyebrow,
   Rest,
+  Gradient,
 };