+++ packages/vital-flowcontainer/src/components/FlowContainer/tokens/vitalFlowContainer.ts
@@ -17,18 +17,27 @@ import {
   addProps, replaceWith, as, on, flowHoc,
 } from '@bodiless/fclasses';
 import {
-  withMandatoryCategories, withAllTitlesFromTerms, ifComponentSelector,
+  withMandatoryCategories,
+  withAllTitlesFromTerms,
+  ifComponentSelector,
+  withLibraryComponents,
 } from '@bodiless/layouts';
 import { asFluidToken, vitalSpacing } from '@bodiless/vital-elements';
 import { vitalImageFlowContainer } from '@bodiless/vital-image';
 import { vitalEditorsFlowContainer } from '@bodiless/vital-editors';
+import { vitalYouTubeFlowContainer } from '@bodiless/vital-youtube';
 import { vitalListFlowContainer } from '@bodiless/vital-list';
+import { vitalCardFlowContainer } from '@bodiless/vital-card';
 import { vitalTableFlowContainer } from '@bodiless/vital-table';
+import { vitalButtonFlowContainer } from '@bodiless/vital-buttons';
 
 import FlowContainerClean, { FlowContainerPreview } from '../FlowContainerClean';
 
+const fluidToken = asFluidToken();
+type FluidToken = typeof fluidToken;
+
 const blacklistCategories = ['Group'];
-const mandatoryCateogries = ['Type'];
+const mandatoryCategories = ['Type'];
 
 /**
  * Token which creates the VitalDS Default Base for a Flow Container
@@ -37,7 +46,7 @@ const Base = asFluidToken({
   Core: {
     _: as(
       withAllTitlesFromTerms({ blacklistCategories }),
-      withMandatoryCategories(mandatoryCateogries),
+      withMandatoryCategories(mandatoryCategories),
       addProps({ blacklistCategories }),
     ),
     // @todo restore tools and component id badge
@@ -106,8 +115,11 @@ const AsFlowContainerItem = asFluidToken({
 const WithBaseVariations = asFluidToken(
   vitalImageFlowContainer.WithImageVariations,
   vitalEditorsFlowContainer.WithEditorVariations,
+  vitalYouTubeFlowContainer.WithYouTubeVariations,
   vitalListFlowContainer.WithListVariations,
+  vitalCardFlowContainer.WithCardVariations,
   vitalTableFlowContainer.WithTableVariations,
+  vitalButtonFlowContainer.WithButtonVariations,
 );
 
 /**
@@ -128,9 +140,6 @@ const ContentRegion = asFluidToken(
   },
 );
 
-/**
- * Adds a content region to a flow container.
- */
 const WithContentRegionVariations = asFluidToken({
   Components: {
     ContentRegion: on(FlowContainerClean)(ContentRegion),
@@ -159,14 +168,107 @@ const Hero = asFluidToken(
   WithBaseVariations,
 );
 
-export default {
+const WithContentLibrary = asFluidToken(
+  {
+    ...Default,
+    Core: {
+      _: withLibraryComponents()
+    },
+  },
+);
+
+/**
+ * Tokens for the vital flow container
+ *
+ * @category Token Collection
+ * @see [[FlowContainerClean]]
+ */
+export interface VitalFlowContainer {
+  Base: FluidToken,
+  /**
+   * Defines the default flow container for the Vital DS.
+   * - Core domain defines constraints on categories.
+   * - Spacing domain defines gutters
+   * - Components domain adds the following basic Vital components:
+   *   - Images, Editors, Lists, Content Region (nested flow container).
+   *
+   * #### Customizing:
+   *
+   * @example Add a component
+   * ```js
+   * import { vitalFlowContainerBase } from '@bodiless/vital-flowcontainer';
+   *
+   * const Default = asFluidToken(vitalFlowContainerBase.Default, {
+   *   Components: {
+   *     MyComponent: on(MyComponentClean)(myComponent.Default),
+   *   }
+   * });
+   * ```
+   */
+  Default: FluidToken,
+  /**
+   * @deprecated
+   * Flow container which can be used in the Hero slot.
+   */
+  Hero: FluidToken,
+  /**
+   * Defins a flow container which is to be used as a content region (that is,
+   * nested within another flow container). This contains all components
+   * defined in the `Default` flow container with the exception of
+   * the Content Region itself (i.e. you can't have double nesting).
+   *
+   * You can shadow this token to change the components which are available
+   * in a content region.  For example:
+   * ```ts
+   * const ContentRegion = asFluidToken({
+   *    ...vitalContentRegionBase.ContentRegion,
+   *    Components: {
+   *      ...vitalContentRegionBase.ContentRegion.Components,
+   *      SomethingNew: on(MyComponentClean)(myComponent.Default),
+   *   },
+   * });
+   * ```
+   */
+  ContentRegion: FluidToken,
+  /**
+   * Composable token which enables a flow container to be nested inside another.
+   */
+  AsFlowContainerItem: FluidToken,
+  /**
+   * Composable token which constrains all items to full width.
+   */
+  WithFullWidthConstraint: FluidToken,
+  /**
+   * Composable token which constrains all items to 1/3 width on tablet.
+   */
+  WithTabletOneThirdConstraint: FluidToken,
+  /**
+   * Allows only a single item in the flow container.
+   */
+  WithSingleConstraint: FluidToken,
+  /**
+   * Composable token which adds content library functionality.
+   */
+  WithContentLibrary: FluidToken,
+}
+
+/**
+ * Tokens for flow containers.
+ *
+ * @category Token Collection
+ * @see [[VitalFlowContainer]]
+ * @see [[FlowContainerClean]]
+ */
+const vitalFlowContainer: VitalFlowContainer = {
   Base,
   Default,
   Hero,
   ContentRegion,
-  WithContentRegionVariations,
   AsFlowContainerItem,
   WithFullWidthConstraint,
   WithTabletOneThirdConstraint,
   WithSingleConstraint,
+  WithContentLibrary,
 };
+
+export default vitalFlowContainer;