+++ packages/vital-layout/src/components/Header/types.ts
@@ -35,6 +35,7 @@ export type HeaderComponents = {
   WhereToBuyWrapper: ComponentOrTag<any>,
   WhereToBuy: ComponentOrTag<any>,
   DesktopSearch: ComponentOrTag<any>,
+  MobileSearch: ComponentOrTag<any>,
   SearchToggler: ComponentOrTag<any>,
 };