+++ packages/vital-layout/src/components/Layout/tokens/vitalLayout.ts
@@ -13,13 +13,14 @@
  */
 
 import { withNodeKey } from '@bodiless/core';
-import { useIsBurgerMenuHidden, withBurgerMenuProvider } from '@bodiless/vital-navigation';
+import { useIsBurgerMenuHidden, withBurgerMenuProvider, withBreadcrumbStore } from '@bodiless/vital-navigation';
 import {
   addProps,
   as,
   flowIf,
   not,
 } from '@bodiless/fclasses';
+import { WithStructuredDataProvider } from '@bodiless/schema-org';
 import { asLayoutToken } from '../LayoutClean';
 import { vitalFooter } from '../../Footer';
 import { vitalHeader } from '../../Header';
@@ -28,14 +29,17 @@ import { LayoutIds } from './constants';
 import { StyleGuide } from './StyleGuide';
 
 /**
-  * Token that defines a basic layout.
-  */
+ * Token that defines a basic layout.
+ */
 const Base = asLayoutToken({
   Core: {
-    _: withBurgerMenuProvider,
+    _: as(withBurgerMenuProvider, withBreadcrumbStore),
   },
   Components: {
     Helmet: vitalHelmet.Default,
+  },
+  SEO: {
+    _: WithStructuredDataProvider,
   },
   Behavior: {
     Container: addProps({ id: LayoutIds.Content }),