+++ packages/vital-templates/src/components/GenericTemplate/tokens/vitalGenericTemplate.ts
@@ -15,17 +15,23 @@
 import {
   on,
   as,
-  Img,
   addProps,
   replaceWith,
   Fragment,
   flowIf,
+  withDesign,
+  Img,
 } from '@bodiless/fclasses';
-import { vitalImage } from '@bodiless/vital-image';
+import { asBodilessChameleon } from '@bodiless/components';
 import { LayoutClean, vitalLayout } from '@bodiless/vital-layout';
 import { vitalFlowContainer } from '@bodiless/vital-flowcontainer';
-import { withNodeKey, useNode } from '@bodiless/core';
+import { withNode, withNodeKey, useNode } from '@bodiless/core';
 import { vitalSpacing, vitalTypography } from '@bodiless/vital-elements';
+import { SearchLayoutClean, vitalSearchLayout } from '@bodiless/vital-search';
+import { vitalBreadcrumbs } from '@bodiless/vital-navigation';
+import { vitalImage } from '@bodiless/vital-image';
+import { YouTubeClean, vitalYouTube } from '@bodiless/vital-youtube';
+import { CardClean, vitalCard } from '@bodiless/vital-card';
 import { asGenericTemplateToken } from '../GenericTemplateClean';
 import { GenericTemplateNodeKeys } from '../constants';
 
@@ -39,28 +45,41 @@ const WithNoBreadcrumbsOnHomePage = asGenericTemplateToken({
   },
 });
 
+const heroDefaultData = {
+  component: 'Image',
+};
+
+const heroUseOverrides = () => ({
+  groupLabel: 'Hero'
+});
+
 const Base = asGenericTemplateToken({
   Components: {
     PageWrapper: on(LayoutClean)(vitalLayout.Default),
-    // @todo breadcrumb placeholder
-    Breadcrumb: addProps({ children: 'Breadcrumb Placeholder', }),
-    // @todo in Hero ticket is change this to chameleon.
-    TopContent: on(Img)(vitalImage.Default, vitalImage.WithLandscapePlaceholder),
+    Breadcrumb: as(vitalBreadcrumbs.Default),
+    TopContent: as(
+      asBodilessChameleon('component', heroDefaultData, heroUseOverrides),
+      withDesign({
+        Image: on(Img)(vitalImage.Hero),
+        Video: on(YouTubeClean)(vitalYouTube.Hero),
+        HeroCard: on(CardClean)(vitalCard.Hero),
+      }),
+    ),
     Content: as(vitalFlowContainer.Default),
     BottomContent: as(vitalFlowContainer.Default),
   },
   Schema: {
-    TopContent: withNodeKey(GenericTemplateNodeKeys.TopContent),
+    TopContent: as(withNode, withNodeKey(GenericTemplateNodeKeys.TopContent)),
     Content: withNodeKey(GenericTemplateNodeKeys.Content),
     BottomContent: withNodeKey(GenericTemplateNodeKeys.BottomContent),
   },
   Spacing: {
-    TopContent: vitalSpacing.WithSiteXLConstraint,
     BreadcrumbWrapper: as(
       vitalSpacing.WithSiteMargin,
       vitalSpacing.WithSiteXLConstraint,
       'my-2.5',
     ),
+    TopWrapper: vitalSpacing.GutterBottom,
     // @todo move styling of breadcrumb to breadcrumb component when it exists.
     Breadcrumb: vitalTypography.Rest,
     ContentWrapper: as(
@@ -81,7 +100,17 @@ const Default = asGenericTemplateToken({
   ...Base,
 });
 
+const Search = asGenericTemplateToken(Default, {
+  Components: {
+    Breadcrumb: addProps({ children: 'Search', }),
+    TopContent: replaceWith(Fragment),
+    Content: on(SearchLayoutClean)(vitalSearchLayout.Default),
+    BottomContent: replaceWith(Fragment),
+  }
+});
+
 export default {
   Base,
   Default,
+  Search,
 };